used promise syntax with await to s.t wrong cx id or pd id does not crash api
